Split off some test functions in other py files
