+++ .gitignore
@@ -8,7 +8,9 @@ tests/data/history
 tests/data/saved
 tests/out.log.geth
 
-
 # Ignore created javascript files
 tests/deploy.js
 tests/fund.js
+
+# python metadata in tests
+tests/*.pyc
